+++ semana17/aula48/template-filtros-ordenacao-paginacao/template-filtros-ordenacao-paginacao/src/types/queryEndpoint.ts
@@ -1,0 +1,17 @@
+import {connection} from ".."
+import {searchRecipeInput} from "../types/searchRecipeInput"
+import { Request, Response } from "express";
+
+
+export default async function selectAllUsers(userName: string):Promise<any> {
+
+
+
+ const result = await connection.raw(`
+      SELECT id, name, email, type
+      FROM aula48_exercicio;
+     WHERE name LIKE "${userName}";
+   `)
+
+   return result[0]
+}
